Checkstyle/PMD fix
