+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/Graph.java
@@ -4,50 +4,46 @@ import java.util.ArrayList;
 
 /**
  * Mainly a dataholder class which represents a graph.
- * @author Cas
  *
+ * @author Cas
  */
 public class Graph {
-	
-	private Node root;
-	private ArrayList<Bubble> nodes;
-	
-	/**
-	 * Constructs a graph from an arrayList of nodes. (the edges are already in these nodes)
-	 * @param nodes
-	 */
-	public Graph(ArrayList<Bubble> nodes) {
-		//this.root = nodes.get(1);
-		this.nodes = nodes;
-	}
-	
-	public void print() {
-		for(Bubble n : nodes) {
-			if (n.getId() != 0)
-				System.out.println(n);
-		}
-	}
 
-	public Bubble getRoot() {
-		return nodes.get(1);
-	}
+  private ArrayList<Bubble> nodes;
 
-	public void setRoot(Node root) {
-		this.root = root;
-	}
+  /**
+   * Constructs a graph from an arrayList of nodes. (the edges are already in these nodes)
+   *
+   * @param nodes all of the nodes of this graph.
+   */
+  public Graph(ArrayList<Bubble> nodes) {
+    this.nodes = nodes;
+  }
 
-	public int getSize() {
-		return nodes.size();
-	}
+  /**
+   * Print a string representation of this node for testing purposes.
+   */
+  public void print() {
+    for (Bubble n : nodes) {
+      if (n.getId() != 0) {
+        System.out.println(n);
+      }
+    }
+  }
 
-	public ArrayList<Bubble> getNodes() {
-		return nodes;
-	}
+  public Bubble getRoot() {
+    return nodes.get(1);
+  }
 
-	public void addNode(Bubble node) {
-		nodes.add(node);
-	}
-	
-	
+  public int getSize() {
+    return nodes.size();
+  }
 
+  public ArrayList<Bubble> getNodes() {
+    return nodes;
+  }
+
+  public void addNode(Bubble node) {
+    nodes.add(node);
+  }
 }
