+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/test/utility/TestAnnotationException.java
@@ -1,0 +1,27 @@
+package nl.tudelft.pl2016gr2.test.utility;
+
+/**
+ * An exception which is thrown when there is something wrong with the test annotation. This
+ * exception is a runtime exception and isn't intended to be caught.
+ *
+ * @see RuntimeException
+ *
+ * @author Faris
+ */
+public class TestAnnotationException extends RuntimeException {
+
+  /**
+   * Creates a new instance of <code>NotAnnotatedException</code> without detail message.
+   */
+  public TestAnnotationException() {
+  }
+
+  /**
+   * Constructs an instance of <code>NotAnnotatedException</code> with the specified detail message.
+   *
+   * @param msg the detail message.
+   */
+  public TestAnnotationException(String msg) {
+    super(msg);
+  }
+}
