+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/test/utility/TestId.java
@@ -1,4 +1,4 @@
-package nl.tudelft.pl2016gr2.testUtility;
+package nl.tudelft.pl2016gr2.test.utility;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,8 +6,8 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation which makes it possible to give private methods and fields an ID
- * so they can be accessed by test classes using the AccessPrivate class.
+ * Annotation which makes it possible to give private methods and fields an ID so they can be
+ * accessed by test classes using the AccessPrivate class.
  *
  * @author Faris
  */
@@ -15,10 +15,10 @@ import java.lang.annotation.Target;
 @Target({ElementType.METHOD, ElementType.FIELD})
 public @interface TestId {
 
-	/**
-	 * The ID of the field or method.
-	 *
-	 * @return the ID.
-	 */
-	public String id();
+  /**
+   * The ID of the field or method.
+   *
+   * @return the ID.
+   */
+  public String id();
 }
